Refactor PSD warning test helpers
Extract transformed full-covariance PSD warning handling into
shared test utilities so transformed and gradient tests use the
same logic.

This reduces duplication and keeps warning expectations scoped to
the tests that exercise PCA or VAE inverse covariance repairs.
